swarm_strands/requirements.txt
strands-agents
chromadb
pymupdf4llm
tensorflow
scikit-learn
nibabel
langchain-docling
langchain-chroma
langchain-community
